split bitbucket-api.js into cloud, dc, and router files
- bitbucket-cloud-api.js: Basic auth + links.diff.href flow for Cloud
- bitbucket-dc-api.js: Bearer auth + JSON→unified diff converter for DC
- bitbucket-api.js: thin router delegating to the right implementation

background.js call site is unchanged (still one import, one call).

Made-with: Cursor
